게시판 - 댓글 기능 추가

+++ src/main/java/com/rest/api/controller/BoardController.java
@@ -1,6 +1,7 @@
 package com.rest.api.controller;
 
 import com.rest.api.domain.Board;
+import com.rest.api.domain.Comment;
 import com.rest.api.domain.User;
 import com.rest.api.repository.BoardQueryRepository;
 import com.rest.api.util.PageRequest;
@@ -38,13 +39,13 @@ public class BoardController {
              @PageableDefault(page = 0, size = 10) PageRequest pageRequest,
              @RequestParam(value="ordr" , defaultValue = "DESC") String ordr,
              @RequestParam(value="ordrNm" , defaultValue = "id") String ordrNm,
-             @RequestParam(value="name" , defaultValue = "") String name,
+             @RequestParam(value="title" , defaultValue = "") String title,
              @RequestParam(value="content" , defaultValue = "") String content , HttpServletRequest req) throws Exception {
         ResponseMessage ms = new ResponseMessage();
-        if (!StringUtils.isEmpty(name) || !StringUtils.isEmpty(content)) {
+        if (!StringUtils.isEmpty(title) || !StringUtils.isEmpty(content)) {
             pageRequest.setPage(1);
         }
-        Page<Board> board = boardQueryRepository.getList(pageRequest.of(ordr,ordrNm) , name , content);
+        Page<Board> board = boardQueryRepository.getList(pageRequest.of(ordr,ordrNm) , title , content);
         ms.add("result",board);
         return ms;
     }
@@ -75,7 +76,52 @@ public class BoardController {
     public ResponseMessage delete(@PathVariable Long id ,Board board , HttpServletRequest req) throws Exception{
         ResponseMessage ms = new ResponseMessage();
         board.setId(id);
-        boardQueryRepository.delete(board);
+
+        if(boardQueryRepository.commentCount(board) == 0){
+            boardQueryRepository.delete(board);
+        }else{
+            ms.setMessage("댓글이 존재하는 게시글은 삭제할 수 없습니다.");
+        }
+        return ms;
+    }
+
+    //@CrossOrigin(origins = "*", allowedHeaders = "*")
+    @GetMapping("/board/{id}/comment")
+    public ResponseMessage  getPostComments(@PathVariable Long id){
+        ResponseMessage ms = new ResponseMessage();
+        Board board = boardQueryRepository.findById(id);
+        ms.add("result",boardQueryRepository.findCommentsByComment(board));
+        return ms;
+    }
+
+    //@CrossOrigin(origins = "*", allowedHeaders = "*")
+    @PostMapping("/board/{id}/comment")
+    public ResponseMessage createComment(@PathVariable Long id, Comment comment){
+        ResponseMessage ms = new ResponseMessage();
+        Board board = boardQueryRepository.findById(id);
+        comment.setBoard(board);
+        boardQueryRepository.commentSave(comment);
+        return ms;
+    }
+
+    //@CrossOrigin(origins = "*", allowedHeaders = "*")
+    @PutMapping("/board/{id}/comment/{commentId}")
+    public ResponseMessage updateComment(@PathVariable Long id, @PathVariable Long commentId, Comment comment){
+        ResponseMessage ms = new ResponseMessage();
+        Board board = boardQueryRepository.findById(id);
+        comment.setComment_id(commentId);
+        comment.setBoard(board);
+        boardQueryRepository.commentUpdate(comment);
+        return ms;
+    }
+
+
+    //@CrossOrigin(origins = "*", allowedHeaders = "*")
+    @DeleteMapping("/board/{id}/comment/{commentId}")
+    public ResponseMessage deleteComment(@PathVariable Long id, @PathVariable Long commentId , Comment comment){
+        ResponseMessage ms = new ResponseMessage();
+        comment.setComment_id(commentId);
+        boardQueryRepository.commentDelete(comment);
         return ms;
     }
